+++ src/components/Theme.ts
@@ -6,20 +6,17 @@ import { INavigationBar } from './NavigationBar/interfaces';
 import { navigationBarDefaultStyles } from './NavigationBar/styles';
 import { ILoader } from './Loader/interfaces';
 import { loaderDefaultStyles } from './Loader/styles';
-import { IArrow } from './Icons/Arrow/interfaces';
-import { arrowDefaultStyles } from './Icons/Arrow/styles';
 import { ITag } from './Tag/interfaces';
 import { tagDefaultStyles } from './Tag/styles';
-import { ITable } from './Table/interfaces';
-import { tableDefaultStyles } from './Table/styles';
 import colors from '../styles/colors';
+import { tableDefaultStyles } from './Table/constants';
+import { ITable } from './Table/interfaces';
 
 interface ITheme {
   wizardStep: IWizardStep;
   button: IButton;
   navigationBar: INavigationBar;
   loader: ILoader;
-  arrow: IArrow;
   tag: ITag;
   table: ITable;
   colors: {
@@ -32,7 +29,6 @@ export const farmerConnectTheme: ITheme = {
   button: buttonDefaultStyles,
   navigationBar: navigationBarDefaultStyles,
   loader: loaderDefaultStyles,
-  arrow: arrowDefaultStyles,
   tag: tagDefaultStyles,
   table: tableDefaultStyles,
   colors,